pyproject.toml
[project]
name = "kis-omega"
version = "0.1.0"
description = "KIS Open API 기반 모의주문 + 간단 전략 스캐폴드"
requires-python = ">=3.11"
dependencies = [
  "httpx==0.27.2",
  "python-dotenv==1.0.1",
  "pydantic==2.8.2",
  "pandas==2.2.2",
  "rich==13.7.1"
]

[tool.pip]
# nothing

[tool.kisomega]
# 프로젝트 전용 툴 섹션(선택)